refactor: 환승 결과 페이지 폴더 구조 변경
