+++ src/pages/TrainTransferResult/components/TransferLoading/main.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import Loading from 'components/Common/Loading';
 
 function TransferLoading() {
